`colorblind_friendly` module renameb to `cbf`
Reasons:
 - word 'colorblind' may be considered offensive,
 - every time the module was imported `as cbf`.
